Remove redundant commands
