+++ aim/__version__.py
@@ -1,3 +1,3 @@
-VERSION = (2, 3, 0)
+VERSION = (2, 4, 2)
 
 __version__ = '.'.join(map(str, VERSION))
