+++ aim/cli/cli.py
@@ -7,11 +7,7 @@ from aim.cli.configs import *
 from aim.cli.init import commands as init_commands
 from aim.cli.experiment import commands as experiment_commands
 from aim.cli.version import commands as version_commands
-from aim.cli.de import commands as de_commands
-# from aim.cli.status import commands as status_commands
-# from aim.cli.config import commands as config_commands
-# from aim.cli.commit import commands as commit_commands
-# from aim.cli.reset import commands as reset_commands
+from aim.cli.up import commands as up_commands
 
 core._verify_python3_env = lambda: None
 
@@ -31,14 +27,4 @@ cli_entry_point.add_command(init_commands.init, INIT_NAME)
 cli_entry_point.add_command(experiment_commands.exp_entry_point,
                             EXPERIMENT_NAME)
 cli_entry_point.add_command(version_commands.version, VERSION_NAME)
-# cli_entry_point.add_command(commit_commands.commit, COMMIT_NAME)
-# cli_entry_point.add_command(de_commands.de_entry_point, DE_NAME)
-# cli_entry_point.add_command(status_commands.status, STATUS_NAME)
-# cli_entry_point.add_command(reset_commands.reset, RESET_NAME)
-# cli_entry_point.add_command(config_commands.config, CONFIG_NAME)
-
-# Development environment management commands
-cli_entry_point.add_command(de_commands.up, DE_UP_NAME)
-# cli_entry_point.add_command(de_commands.down, DE_DOWN_NAME)
-# cli_entry_point.add_command(de_commands.upgrade, DE_UPGRADE_NAME)
-# cli_entry_point.add_command(de_commands.pull, DE_PULL_NAME)
+cli_entry_point.add_command(up_commands.up, UP_NAME)
