+++ aim/cli/configs.py
@@ -1,14 +1,4 @@
 INIT_NAME = 'init'
 EXPERIMENT_NAME = 'experiment'
-COMMIT_NAME = 'commit'
-RESET_NAME = 'reset'
 VERSION_NAME = 'version'
-DE_NAME = 'de'
-STATUS_NAME = 'status'
-CONFIG_NAME = 'config'
-
-# Shortcuts
-DE_UP_NAME = 'up'
-DE_DOWN_NAME = 'down'
-DE_UPGRADE_NAME = 'upgrade'
-DE_PULL_NAME = 'pull'
+UP_NAME = 'up'
