+++ aim/engine/configs.py
@@ -40,23 +40,25 @@ AIM_PROFILE_NAME = '.aim_profile'
 AIM_PROFILE_CONFIG_FILE_NAME = 'config.json'
 AIM_PROFILE_SSH_DIR_NAME = '.ssh'
 
-# Aim Container
-AIM_CONTAINER_IMAGE_NAME = 'aimhubio/aim-board'
-AIM_CONTAINER_IMAGE_DEFAULT_TAG = 'latest'
-AIM_CONTAINER_IMAGE_DEV = 'aimhub-live'
-AIM_CONTAINER_PREFIX = 'aim_container'
-AIM_CONTAINER_TELEMETRY_FLAG = 'AIM_UI_TELEMETRY_ENABLED'
-
 # Aim Resource Collector
 AIM_RESOURCE_METRIC_PREFIX = '__system__'
 DEFAULT_SYSTEM_TRACKING_INT = 10
 
-# Web
-AIM_CONTAINER_DEFAULT_PORT = 43800
-AIM_CONTAINER_DEFAULT_HOST = '127.0.0.1'
-AIM_CONTAINER_CMD_PORT = 43803
+# Aim UI
+AIM_UI_DEFAULT_PORT = 43800
+AIM_UI_DEFAULT_HOST = '127.0.0.1'
+AIM_TF_LOGS_PATH_KEY = '__AIM_TF_LOGS_PATH__'
+AIM_FLASK_ENV_KEY = '__AIM_FLASK_ENV__'
+AIM_UI_TELEMETRY_KEY = 'AIM_UI_TELEMETRY_ENABLED'
+
+# Aim Container (deprecated, will be removed in the next versions)
+AIM_CONTAINER_IMAGE_NAME = 'aimhubio/aim-board'
+AIM_CONTAINER_IMAGE_DEFAULT_TAG = 'latest'
+AIM_CONTAINER_IMAGE_DEV = 'aimhub-live'
+AIM_CONTAINER_PREFIX = 'aim_container'
+
+# UI processes (deprecated, will be removed in the next versions)
 AIM_PROCESS_ENV_VAR = '__AIM_PROCESS_UUID__'
 AIM_BRANCH_ENV_VAR = '__AIM_BRANCH__'
 AIM_COMMIT_ENV_VAR = '__AIM_COMMIT__'
 AIM_AUTOMATED_EXEC_ENV_VAR = '__AIM_AUTOMATED_EXEC__'
-AIM_TF_LOGS_PATH = '/tf_logs'
